refactor(components): Add index and Pascal case convention
