+++ guifaffinity-front/src/components/gif/index.ts
@@ -1,0 +1,1 @@
+export { Gif } from "./Gif";
