+++ guifaffinity-front/src/components/header/header.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import { Header } from "./header";
+import { Header } from "./Header";
 
 test("display logo", () => {
   render(<Header />);
