+++ guifaffinity-front/src/components/header/index.ts
@@ -1,0 +1,1 @@
+import { Header } from "./Header";
