+++ guifaffinity-front/src/components/search/index.ts
@@ -1,0 +1,1 @@
+export { Search } from "./Search";
